+++ src/lxc/Makefile.am
@@ -9,6 +9,7 @@ noinst_HEADERS = \
 	bdev/bdev.h \
 	bdev/lxcaufs.h \
 	bdev/lxcbtrfs.h \
+	bdev/lxcdir.h \
 	bdev/lxclvm.h \
 	bdev/lxcoverlay.h \
 	bdev/lxcrsync.h \
@@ -68,6 +69,7 @@ liblxc_so_SOURCES = \
 	bdev/bdev.c bdev/bdev.h \
 	bdev/lxcaufs.c bdev/lxcaufs.h \
 	bdev/lxcbtrfs.c bdev/lxcbtrfs.h \
+	bdev/lxcdir.c bdev/lxcdir.h \
 	bdev/lxclvm.c bdev/lxclvm.h \
 	bdev/lxcoverlay.c bdev/lxcoverlay.h \
 	bdev/lxcrsync.c bdev/lxcrsync.h \
